tile descriptions: consolidate to single string for 'Industrial'
Formerly #23 and #12 were both "Industrial", with #12 being used for
the basic industrial zones, and #23 being used for any animation
appearing for industrial zones.

Now that the tiles using a particular description number don't need
to be contiguous, we will consolidate to just use #12 for "industrial",
and this will free up #23 for some other use in the future.

git-svn-id: https://micropolis.googlecode.com/svn/trunk/micropolis-java@827 d9718cc8-9f43-0410-858b-315f434eb58c