traffic: perimeter-discount should handle partly-deconstructed zones
if a 3x3 zone was partly bulldozed and a road was inserted,
do not count that road tile as part of the zone for the purposes
of finding the zone perimeter. I.e. the sim still has to pay the cost
of walking from the road tile onto a zone tile.