bug fixes and completions written by abelian_determinants
>>>
Here is my modified version of DiverCity.  I basically made bug-fixes and completions, the biggest of which was adding and extending your XML save-file format to include the new features in DiverCity.  I also made a slight tweak to the traffic simulation code to reduce the number of A* runs for larger cities.

Daniel
<<<